feature: ribosome database building

+++ bin/ribo_db_build.py
@@ -2,6 +2,19 @@ import re, os
 from dataclasses import dataclass
 from typing import List, Dict, Tuple, Generator
 
+def progressbar(iteration, total, prefix = '', suffix = '', filler = '█', printEnd = "\r") -> None:
+    """
+    Show a progress bar indicating downloading progress
+    """
+    percent = f'{round(100 * (iteration / float(total)), 1)}'
+
+    add = int(100 * iteration // total)
+    bar = filler * add + '-' * (100 - add)
+    print(f'\r{prefix} |{bar}| {percent}% {suffix}', end = printEnd)
+
+    if iteration == total: 
+        print()
+
 @dataclass
 class GenBankElement:
     organism_name : str
@@ -12,13 +25,18 @@ class GenBankElement:
 def load_files(ribo_db_dir : str) -> Generator:
     """
     Load one after the other the GenBank general files to parse.
+    Remove GenBank .gb files after use.
     """
-    for file in os.listdir(ribo_db_dir):
+    size = len(os.listdir(ribo_db_dir))
+    for index, file in enumerate(os.listdir(ribo_db_dir)):
+        progressbar(index+1, size)
         if file.endswith('.gb'):
             filepath = os.path.join(ribo_db_dir, file)
+
             with open(filepath, 'r', encoding='utf-8') as ribo_file:
                 yield ribo_file.read()
-        os.remove(filepath)
+
+            os.remove(filepath)
 
 def postprocess_regex(pattern : List[Tuple[str]]) -> List[List[str]]:
     liste = []
@@ -34,18 +52,20 @@ def parser(ribo_db_dir : str):
     Uses RegEx for parsing : sensible to GenBank Summary format.
     """
     for ribo_file_content in load_files(ribo_db_dir):
-        genome_seq = ''.join(re.findall('[atcg]+', ribo_file_content.split('ORIGIN')[1]))
+        genome_seq = ''.join(re.findall('[atcg]+', ribo_file_content))
         accession_id = re.findall(r'VERSION     (.*?(?=\n))', ribo_file_content)[0]
         organism_name = re.findall(r'ORGANISM..(.+?)(?=\n)', ribo_file_content)[0]
-        pattern = re.findall(r'(?s)rRNA\s+(?:complement\(join\(([\d\.\.,\s]+)\)\)|complement\((\d+\.\.\d+)\)|(\d+\.\.\d+)).*?product=\"(.*?)(?= ribosomal RNA)',
-                             ribo_file_content)
+            
+        pattern = re.findall(r'rRNA\s+(?:complement\(join\(([\d\.\.,\s]+)\)\)|complement\((\d+\.\.\d+)\)|(\d+\.\.\d+))\s+.*\s+.*\s+.*?(?<=product=\")(.+?)(?=\")',
+                        ribo_file_content)
         pattern = postprocess_regex(pattern)
-        ribo_dico = {}
+        ribo_dico = {elt[1] : [] for elt in pattern}
+
         for elt in pattern:
             beacons, ribo_id = elt
             beacons_list = re.findall('([0-9]+..[0-9]+)', beacons)
 
-            ribo_dico[ribo_id] = [(int(beacons.split('..')[0]), int(beacons.split('..')[1])) for beacons in beacons_list]
+            ribo_dico[ribo_id].extend([(int(beacons.split('..')[0]), int(beacons.split('..')[1])) for beacons in beacons_list])
         
         ribo_seq_dico = {ribo_id : [genome_seq[beacon[0]:beacon[1]] for beacon in beacons] for ribo_id, beacons in ribo_dico.items()}      
         
@@ -57,8 +77,15 @@ def parser(ribo_db_dir : str):
             )
 
 def prepare_ribo_db(ribo_db_dir : str):
+    """
+    Write ribosomic sequences to a new fasta file.
+    """
+    n = 0
     for ribo_elts in parser(ribo_db_dir):
-        name_file = f'ribosomes_{ribo_elts.organism_name}.fasta'
+        if ribo_elts == "Timeout":
+            continue
+        name = ribo_elts.organism_name.split('/')[0]
+        name_file = f'ribosomes_{name}.fasta'
         out_path = os.path.join(ribo_db_dir, name_file)
         if name_file in os.listdir(ribo_db_dir):
             os.remove(out_path)
@@ -69,8 +96,10 @@ def prepare_ribo_db(ribo_db_dir : str):
                     ribofile.write(
                         f">{ribo_id}_{index}_{ribo_elts.organism_name}|{beacons[index][0]}-{beacons[index][1]}\n{seq.upper()}\n"
                     )
-    return 0
+        n+=1
+    
+    print("\n   Ribosome database completed\n")
+    return n
             
-
-
+#prepare_ribo_db(r"C:\Subpbiotech_cours\BT5\BIM_BMC\GENOM\project\project_git\GENOM\db\ribo_db")
